Updated the Readme (see details!)
I changed the README.md to match THIS repository and to provide installation instructions for THIS repository, as well as adding links to get back to the other two forked repositories in case the user wants to be on one of those (got here by accident). Also added instructions for Windows, removed the '$' in the instructions so users can mark-copy-paste all lines at once. Also updated the instructions to apply to versions 4.0 and 4.5.X, which at the time of this writing, were released since 3.0 and up to this point.